backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data processing
pandas==2.2.0
openpyxl==3.1.2
numpy

# PDF processing
pdfplumber==0.10.3
pymupdf

# Image handling & OCR
Pillow==10.2.0
opencv-python-headless
paddlepaddle
paddleocr

# Async HTTP (for future AI API integration)
httpx==0.26.0

# AI API Integration
groq==0.4.2
python-dotenv==1.0.0

# Authentication & Database
motor==3.7.1
python-jose[cryptography]==3.3.0
bcrypt
email-validator
